updated and created code for 16-20OCt

+++ 13Oct2023/Lab024AreaOfCircle.py
@@ -1,3 +1,3 @@
-r = int(input("Enter the radius of circle \n"))
+r = float(input("Enter the radius of circle \n"))
 area = 3.14 * r**2
 print("the area of circle is : %.2f"%area)
